requirements.txt
ipykernel
pandas
numpy
scikit-learn
catboost
xgboost
kaggle
matplotlib
streamlit
pyarrow==15
notebook
seaborn
scipy
joblib
python-box==5.4.1
pyyaml
types-PyYAML
tqdm
ensure==1.0.2
lightgbm
imbalanced-learn

-e .
